# make delete faster, use an NIO file walker