seems almost twice as fast in windows

```
1455 File(s) copied
1455 File(s) copied
1455 File(s) copied
old delete 1 took 1849.766 ms
1455 File(s) copied
new delete 1 took 1045.31 ms
1455 File(s) copied
old delete 2 took 1860.53 ms
1455 File(s) copied
new delete 2 took 1044.366 ms
1455 File(s) copied
old delete 3 took 1875.247 ms
1455 File(s) copied
new delete 3 took 1056.219 ms
1455 File(s) copied
old delete 4 took 1833.669 ms
1455 File(s) copied
new delete 4 took 1008.448 ms
1455 File(s) copied
old delete 5 took 1837.931 ms
1455 File(s) copied
new delete 5 took 1029.957 ms
```