@eed3si9n 2018-02-22
Seems to be missing some import.

```
[error] /home/travis/build/sbt/io/io/src/main/scala/sbt/io/IO.scala:459:48: not found: type BasicFileAttributes
[error]       override def visitFile(file: Path, attr: BasicFileAttributes): FileVisitResult = {
[error]                                                ^
```