@eed3si9n 2018-02-22
Travis seems to be confused.
I am going to close and re-open the PR.