@eed3si9n 2018-02-22
Unless this is some blocker, my vote would be to target this to 1.x (and ship it as part of 1.2.0). Would that be ok?