@dwijnand 2018-02-22
> I am happy to have this retargeted to whatever build line you prefer.
> Not sure if the question was aimed to me though :-)

Ok, cool, changing the target branch to 1.x then.