@typesafe-tools 2018-02-22
[//]: # (Refs: 1.x-1.x-pull/133/head-1.x-1.x-1.x)
The validator <A HREF="https://jenkins.scala-sbt.org:8592/job/sbt-validator/489/">has checked</A> the following projects against Scala 2.12,
tested using dbuild, projects built on top of each other.

Project             | Reference            | Commit
:------------------ | :------------------- | :----------------
  sbt               | <A HREF="https://github.com/sbt/sbt/tree/1.x">1.x                 </A> | sbt/sbt@e078c126d15a1967910357a1a2e06971a56c66c6
  zinc              | <A HREF="https://github.com/sbt/zinc/tree/1.x">1.x                 </A> | sbt/zinc@ad46221ca148629df111dbc1561ac348d2b15c9c
  io                | <A HREF="https://github.com/sbt/io/pull/133">pull/133/head       </A> | sbt/io@ce3dc9865e8c688678f6d8b17e4c603cbcdb9877
  librarymanagement | <A HREF="https://github.com/sbt/librarymanagement/tree/1.x">1.x                 </A> | sbt/librarymanagement@0d21ae6369a3e4054ed434902ea84b2520d21da9
  util              | <A HREF="https://github.com/sbt/util/tree/1.x">1.x                 </A> | sbt/util@85f7d807e116b0247c233259f66b592e05a14ffe
  website           | <A HREF="https://github.com/sbt/website/tree/1.x">1.x                 </A> | 

:x: The result is: ***FAILED***
<A HREF="https://jenkins.scala-sbt.org:8592/job/sbt-validator/parambuild?sbt=1.x&zinc=1.x&io=pull/133/head&librarymanagement=1.x&util=1.x&website=1.x&deploy_to_bintray=false&run_sbt_community=false">(restart)</A>