@cunei 2018-03-04
The cause for the test failures is a difference in semantics between the old code and the proposed one. In particular, the previous delete code does not complain when attempting to delete non-existing files or directories, but just ignores them. I prepared a patched pull request that restores the original behavior at https://github.com/sbt/io/pull/135