@mkeskells 2018-03-04
How does the code requite to delete a non existent directory? 
Is this just a race condition or a legitimate detectable case?
If something is changing the files system during a delete should it fail fast or not. I don't know the semantics that are required here, but it seems strange that files are not present, or that it should be ignored