@typesafe-tools 2018-03-12
A validation involving this pull request <a href="https://jenkins.scala-sbt.org:8592/job/sbt-validator/538/console">is in progress...</a>